Implement MembershipJdbcDao
It delegates to stored procedure 
WEBAPP.CustomerValidateMembership.
Implementation was lifted as-is from the old 
taxameter app and unit test written around it to
ensure functionality is maintained.
Had to replace H2 with HSQLDB because H2 didn’t
support OUT parameters in stored procedures.
We now have the way open for refactoring.