Update travelcard conversion script to the one used in production
Previously it was the one used in NL-ACC, which
didn’t actually talk to an SFTP server.
This commit also incorporates a change introduced
by Max whereby the converted csv file is written
to in a “work” directory before being moved to the
target directory. This is done to avoid having
the file picked from the target directory before
all the writing is done.