Fix issue where bean binding exceptions would interrupt validator #2036
As it turns out the validator is able to 
validate more fields than the Errors target has
access to. This happens because the validator acts
on a Customer instance but may bind errors to an 
object that has a subset of the fields that Customer
has. The contract between the Errors and Customer
is untyped anyway (it’s based on the field name) so
we might as well wrap the errors in an unsafe proxy
that gracefully ignores binding errors.