Give up bean binding on account activation validations
Because these validators bind against different
objects: forms on the Admin UI and just the
customer object on the account activation task.
Introducing a map-backed UnboundErrors class
lets us forego bean binding, instead the binding
is done implicitly (by name) on the controller,
where form field names should match the field names
used in the validators.
This also lets us get rid of the previously used
“unsafe proxy” on ClassValidator.
Furthermore a small AccountValidationService is
introduced to force the use of the UnboundErrors