Require obtaining distributed lock when trying to activate customer
Also reload the customer instance once lock is
obtained.
A lock is NOT the ideal solution but is the 
cheapest one since it doesn’t require changing the
UI. Ideally customer activation would run async
by means of a message queue, however currently
UI relies on synchronous confirmation of the 
activation.